+++ packages/browser/src/integrations/breadcrumbs.ts
@@ -14,7 +14,7 @@ import {
 /** JSDoc */
 interface BreadcrumbsOptions {
   console: boolean;
-  dom: boolean | { serializeAttribute: string };
+  dom: boolean | { serializeAttribute: string | string[] };
   fetch: boolean;
   history: boolean;
   sentry: boolean;
@@ -162,13 +162,17 @@ export class Breadcrumbs implements Integration {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   private _domBreadcrumb(handlerData: { [key: string]: any }): void {
     let target;
-    const keyAttr = typeof this._options.dom === 'object' ? this._options.dom.serializeAttribute : undefined;
+    let keyAttrs = typeof this._options.dom === 'object' ? this._options.dom.serializeAttribute : undefined;
+
+    if (typeof keyAttrs === 'string') {
+      keyAttrs = [keyAttrs];
+    }
 
     // Accessing event.target can throw (see getsentry/raven-js#838, #768)
     try {
       target = handlerData.event.target
-        ? htmlTreeAsString(handlerData.event.target as Node, keyAttr)
-        : htmlTreeAsString((handlerData.event as unknown) as Node, keyAttr);
+        ? htmlTreeAsString(handlerData.event.target as Node, keyAttrs)
+        : htmlTreeAsString((handlerData.event as unknown) as Node, keyAttrs);
     } catch (e) {
       target = '<unknown>';
     }