+++ packages/nextjs/src/config/index.ts
@@ -11,23 +11,19 @@ import { constructWebpackConfigFunction } from './webpack';
 export function withSentryConfig(
   userNextConfig: ExportedNextConfig = {},
   userSentryWebpackPluginOptions: Partial<SentryWebpackPluginOptions> = {},
-): NextConfigFunction {
-  const newWebpackExport = constructWebpackConfigFunction(userNextConfig, userSentryWebpackPluginOptions);
-
-  const finalNextConfig = (
-    phase: string,
-    defaults: { defaultConfig: { [key: string]: unknown } },
-  ): NextConfigObject => {
-    const materializedUserNextConfig =
-      typeof userNextConfig === 'function' ? userNextConfig(phase, defaults) : userNextConfig;
-
-    return {
-      ...materializedUserNextConfig,
-      // TODO When we add a way to disable the webpack plugin, doing so should turn this off, too
-      productionBrowserSourceMaps: true,
-      webpack: newWebpackExport,
+): NextConfigFunction | NextConfigObject {
+  // If the user has passed us a function, we need to return a function, so that we have access to `phase` and
+  // `defaults` in order to pass them along to the user's function
+  if (typeof userNextConfig === 'function') {
+    return function(phase: string, defaults: { defaultConfig: { [key: string]: unknown } }): NextConfigObject {
+      const materializedUserNextConfig = userNextConfig(phase, defaults);
+      return {
+        ...materializedUserNextConfig,
+        webpack: constructWebpackConfigFunction(materializedUserNextConfig, userSentryWebpackPluginOptions),
+      };
     };
-  };
+  }
 
-  return finalNextConfig;
+  // Otherwise, we can just merge their config with ours and return an object.
+  return { ...userNextConfig, webpack: constructWebpackConfigFunction(userNextConfig, userSentryWebpackPluginOptions) };
 }