+++ packages/nextjs/test/integration/next.config.js
@@ -1,6 +1,10 @@
 const { withSentryConfig } = require('@sentry/nextjs');
 
-const moduleExports = {};
+const moduleExports = {
+  eslint: {
+    ignoreDuringBuilds: true,
+  }
+};
 const SentryWebpackPluginOptions = {
   dryRun: true,
   silent: true,